feat : "add place admin page"
